@@ -1,46 +1,52 @@
 {
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/server.js",
-  "scripts": {
-    "dev": "nodemon src/server.ts",
-    "build": "tsc --build",
-    "start": "node dist/src/server.js",
-    "clean": "tsc --build --clean",
-    "lint": "eslint --ext .ts ."
-  },
-  "keywords": [],
-  "author": "",
-  "license": "GPL-3.0",
-  "devDependencies": {
-    "@types/bcrypt": "^5.0.0",
-    "@types/express": "^4.17.17",
-    "@types/jsonwebtoken": "^9.0.1",
-    "@types/node": "^18.15.11",
-    "@types/nodemailer": "^6.4.7",
-    "@typescript-eslint/eslint-plugin": "^5.52.0",
-    "eslint": "^8.34.0",
-    "eslint-config-standard-with-typescript": "^34.0.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-n": "^15.6.1",
-    "eslint-plugin-promise": "^6.1.1",
-    "prisma": "^4.11.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.5"
-  },
-  "dependencies": {
-    "@prisma/client": "^4.11.0",
-    "@types/http-server": "^0.12.1",
-    "@types/node-fetch": "^2.6.2",
-    "bcrypt": "^5.1.0",
-    "express": "^4.18.2",
-    "isemail": "^3.2.0",
-    "jsonwebtoken": "^9.0.0",
-    "mariadb": "^3.0.2",
-    "node-fetch": "^2.6.1",
-    "nodemailer": "^6.9.1",
-    "redoc": "^2.0.0",
-    "stripe": "^14.5.0"
-  }
+    "name": "api",
+    "version": "1.0.0",
+    "description": "",
+    "type": "module",
+    "main": "index.js",
+    "scripts": {
+        "dev": "nodemon",
+        "build": "tsup src/index.ts --format esm --clean",
+        "docs": "ts-node src/swagger.mjs",
+        "start": "node dist/index.js",
+        "generate": "prisma generate",
+        "migrate-dev": "prisma migrate dev --name init",
+        "migrate-deploy": "prisma migrate deploy"
+    },
+    "author": {
+        "name": "FurWaz",
+        "email": "contact@furwaz.fr"
+    },
+    "license": "GPL-3.0",
+    "dependencies": {
+        "@prisma/client": "^5.7.1",
+        "and": "^0.0.3",
+        "bcrypt": "^5.1.1",
+        "bee-queue": "^1.7.1",
+        "cors": "^2.8.5",
+        "express": "^4.18.2",
+        "joi": "^17.11.0",
+        "jsonwebtoken": "^9.0.2",
+        "nodemailer": "^6.9.8",
+        "nodemon": "^3.0.2",
+        "prisma": "^5.7.1",
+        "redis": "^4.6.12",
+        "swagger-autogen": "^2.23.7",
+        "swagger-ui-express": "^5.0.0",
+        "ts-node": "^10.9.2",
+        "tsup": "^8.0.1"
+    },
+    "devDependencies": {
+        "@types/bcrypt": "^5.0.2",
+        "@types/cors": "^2.8.17",
+        "@types/express": "^4.17.21",
+        "@types/jsonwebtoken": "^9.0.5",
+        "@types/nodemailer": "^6.4.14",
+        "@types/swagger-ui-express": "^4.1.6",
+        "@typescript-eslint/eslint-plugin": "^6.18.0",
+        "@typescript-eslint/parser": "^6.18.0",
+        "eslint": "^8.56.0",
+        "tsconfig-paths": "^4.2.0",
+        "typescript": "^5.3.3"
+    }
 }
